requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
torch>=2.0
matplotlib>=3.7
seaborn>=0.12
tqdm>=4.65
pyyaml>=6.0
requests>=2.28
fpdf2>=2.7
python-pptx>=0.6.23
